After Successful Synchronizer Testing

+++ bubble_sensors/src/synchronizer.cpp
@@ -14,7 +14,7 @@ data and kalman-filterred orientation data from the real IMU
 
 class Synchronizer : public rclcpp::Node
 {
-  public:
+public:
   //constructor
   Synchronizer()
   : Node("synchronizer")
@@ -55,7 +55,8 @@ class Synchronizer : public rclcpp::Node
       std::bind(&Synchronizer::dvl_sub_callback, this, std::placeholders::_1)
     );
 
-    std::chrono::duration<double> period = std::chrono::duration<double>(1.0 / publishing_frequency_);
+    std::chrono::duration<double> period = std::chrono::duration<double>(1.0 /
+      publishing_frequency_);
     // auto ros_clock = std::make_shared<rclcpp::Clock>(RCL_ROS_TIME);
     timer_ = this->create_timer(
       period,
@@ -63,73 +64,77 @@ class Synchronizer : public rclcpp::Node
     );
 
     imu_orient_pub_ = this->create_publisher<geometry_msgs::msg::PoseWithCovarianceStamped>(
-      imu_orient_input_topic_+"_synced",
+      imu_orient_input_topic_ + "_synced",
       10
     );
 
     dvl_pub_ = this->create_publisher<geometry_msgs::msg::TwistWithCovarianceStamped>(
-      dvl_input_topic_+"_synced",
+      dvl_input_topic_ + "_synced",
       10
     );
 
     imu_accels_pub_ = this->create_publisher<sensor_msgs::msg::Imu>(
-      imu_accels_input_topic_+"_synced",
+      imu_accels_input_topic_ + "_synced",
       10
     );
   }
 
-  private:
-    void imu_accel_sub_callback(const sensor_msgs::msg::Imu &accel_msg){
-      latest_imu_accel_msg_ = accel_msg;
-    }
+private:
+  void imu_accel_sub_callback(const sensor_msgs::msg::Imu & accel_msg)
+  {
+    latest_imu_accel_msg_ = accel_msg;
+  }
 
-    void imu_orient_sub_callback(const geometry_msgs::msg::PoseWithCovarianceStamped &orient_msg){
-      latest_imu_orient_msg_ = orient_msg;
-    }
+  void imu_orient_sub_callback(const geometry_msgs::msg::PoseWithCovarianceStamped & orient_msg)
+  {
+    latest_imu_orient_msg_ = orient_msg;
+  }
 
-    void dvl_sub_callback(const geometry_msgs::msg::TwistWithCovarianceStamped &dvl_msg){
-      latest_dvl_msg_ = dvl_msg;
-    }
+  void dvl_sub_callback(const geometry_msgs::msg::TwistWithCovarianceStamped & dvl_msg)
+  {
+    latest_dvl_msg_ = dvl_msg;
+  }
 
-    void synchronizer_callback(){
-      rclcpp::Time synced_time = this->now();
+  void synchronizer_callback()
+  {
+    rclcpp::Time synced_time = this->now();
 
       // Update header.stamp for each message
-      latest_imu_accel_msg_.header.stamp = synced_time;
-      latest_imu_orient_msg_.header.stamp = synced_time;
-      latest_dvl_msg_.header.stamp = synced_time;
+    latest_imu_accel_msg_.header.stamp = synced_time;
+    latest_imu_orient_msg_.header.stamp = synced_time;
+    latest_dvl_msg_.header.stamp = synced_time;
 
       // Publish the synchronized messages
-      imu_accels_pub_->publish(latest_imu_accel_msg_);
-      imu_orient_pub_->publish(latest_imu_orient_msg_);
-      dvl_pub_->publish(latest_dvl_msg_);
-    }
+    imu_accels_pub_->publish(latest_imu_accel_msg_);
+    imu_orient_pub_->publish(latest_imu_orient_msg_);
+    dvl_pub_->publish(latest_dvl_msg_);
+  }
 
-    float publishing_frequency_;
-    std::string imu_accels_input_topic_;
-    std::string imu_orient_input_topic_;
-    std::string dvl_input_topic_;
+  double publishing_frequency_;
+  std::string imu_accels_input_topic_;
+  std::string imu_orient_input_topic_;
+  std::string dvl_input_topic_;
 
-    rclcpp::Subscription<sensor_msgs::msg::Imu>::SharedPtr imu_accels_sub_;
-    rclcpp::Subscription<geometry_msgs::msg::PoseWithCovarianceStamped>::SharedPtr imu_orient_sub_;
-    rclcpp::Subscription<geometry_msgs::msg::TwistWithCovarianceStamped>::SharedPtr dvl_sub_;
+  rclcpp::Subscription<sensor_msgs::msg::Imu>::SharedPtr imu_accels_sub_;
+  rclcpp::Subscription<geometry_msgs::msg::PoseWithCovarianceStamped>::SharedPtr imu_orient_sub_;
+  rclcpp::Subscription<geometry_msgs::msg::TwistWithCovarianceStamped>::SharedPtr dvl_sub_;
 
-    rclcpp::Publisher<sensor_msgs::msg::Imu>::SharedPtr imu_accels_pub_;
-    rclcpp::Publisher<geometry_msgs::msg::PoseWithCovarianceStamped>::SharedPtr imu_orient_pub_;
-    rclcpp::Publisher<geometry_msgs::msg::TwistWithCovarianceStamped>::SharedPtr dvl_pub_;
+  rclcpp::Publisher<sensor_msgs::msg::Imu>::SharedPtr imu_accels_pub_;
+  rclcpp::Publisher<geometry_msgs::msg::PoseWithCovarianceStamped>::SharedPtr imu_orient_pub_;
+  rclcpp::Publisher<geometry_msgs::msg::TwistWithCovarianceStamped>::SharedPtr dvl_pub_;
 
-    rclcpp::TimerBase::SharedPtr timer_;
+  rclcpp::TimerBase::SharedPtr timer_;
 
-    sensor_msgs::msg::Imu latest_imu_accel_msg_;
-    geometry_msgs::msg::PoseWithCovarianceStamped latest_imu_orient_msg_;
-    geometry_msgs::msg::TwistWithCovarianceStamped latest_dvl_msg_;
+  sensor_msgs::msg::Imu latest_imu_accel_msg_;
+  geometry_msgs::msg::PoseWithCovarianceStamped latest_imu_orient_msg_;
+  geometry_msgs::msg::TwistWithCovarianceStamped latest_dvl_msg_;
 
 };
 
 int main(int argc, char *argv[])
 {
-rclcpp::init(argc, argv);
-rclcpp::spin(std::make_shared<Synchronizer>());
-rclcpp::shutdown();
-return 0;
+  rclcpp::init(argc, argv);
+  rclcpp::spin(std::make_shared<Synchronizer>());
+  rclcpp::shutdown();
+  return 0;
 }
